+++ cli/src/commands/packages.rs
@@ -1,23 +1,21 @@
 use std::str::FromStr;
 
 use ansi_term::Color::Blue;
-use anyhow::anyhow;
 use clap::ArgMatches;
 use phylum_types::types::package::{PackageDescriptor, PackageType};
 use reqwest::StatusCode;
 
 use crate::api::PhylumApi;
 use crate::commands::{CommandResult, ExitCode};
-use crate::print::print_response;
+use crate::filter::{Filter, FilterIssues};
+use crate::format::Format;
 use crate::print_user_warning;
 
-fn parse_package(options: &ArgMatches, request_type: &PackageType) -> Option<PackageDescriptor> {
-    if !(options.is_present("name") && options.is_present("version")) {
-        return None;
-    }
-
-    let name = options.value_of("name").unwrap().to_string(); // required option
+fn parse_package(options: &ArgMatches, request_type: &PackageType) -> PackageDescriptor {
+    // Read required options.
+    let name = options.value_of("name").unwrap().to_string();
     let version = options.value_of("version").unwrap().to_string();
+
     let mut package_type = request_type.to_owned();
 
     // If a package type was provided on the command line, prefer that
@@ -27,27 +25,32 @@ fn parse_package(options: &ArgMatches, request_type: &PackageType) -> Option<Pac
             .unwrap_or(package_type);
     }
 
-    Some(PackageDescriptor { name, version, package_type })
+    PackageDescriptor { name, version, package_type }
 }
 
 /// Handle the subcommands for the `package` subcommand.
 pub async fn handle_get_package(api: &mut PhylumApi, matches: &clap::ArgMatches) -> CommandResult {
     let pretty_print = !matches.is_present("json");
-    let pkg = parse_package(matches, &api.config().request_type);
-    if pkg.is_none() {
-        return Err(anyhow!("Could not find or parse package information"));
-    }
-    let resp = api.get_package_details(&pkg.unwrap()).await;
-    log::debug!("==> {:?}", resp);
 
-    if let Err(Some(StatusCode::NOT_FOUND)) = resp.as_ref().map_err(|e| e.status()) {
-        print_user_warning!(
-            "No matching packages found. Submit a lockfile for processing:\n\n\t{}\n",
-            Blue.paint("phylum analyze <lock_file>")
-        );
-        Ok(ExitCode::PackageNotFound.into())
-    } else {
-        print_response(&resp, pretty_print, None);
-        Ok(ExitCode::Ok.into())
+    let pkg = parse_package(matches, &api.config().request_type);
+    let mut resp = match api.get_package_details(&pkg).await {
+        Ok(resp) => resp,
+        Err(err) if err.status() == Some(StatusCode::NOT_FOUND) => {
+            print_user_warning!(
+                "No matching packages found. Submit a lockfile for processing:\n\n\t{}\n",
+                Blue.paint("phylum analyze <lock_file>")
+            );
+            return Ok(ExitCode::PackageNotFound.into());
+        },
+        Err(err) => return Err(err.into()),
+    };
+
+    let filter = matches.value_of("filter").and_then(|v| Filter::from_str(v).ok());
+    if let Some(filter) = filter {
+        resp.filter(&filter);
     }
+
+    resp.write_stdout(pretty_print);
+
+    Ok(ExitCode::Ok.into())
 }
