+++ cli/src/print.rs
@@ -1,15 +1,9 @@
 use std::borrow::Cow;
-use std::io::{self, Write};
 
 use ansi_term::Color::{Blue, Cyan};
 use clap::Command;
 use prettytable::format;
-use serde::Serialize;
 use unicode_width::{UnicodeWidthChar, UnicodeWidthStr};
-
-use crate::api::PhylumApiError;
-use crate::filter::Filter;
-use crate::summarize::Summarize;
 
 #[macro_export]
 macro_rules! print_user_success {
@@ -33,40 +27,6 @@ macro_rules! print_user_failure {
         eprint!("❗ ");
         eprintln!($($tts)*);
     }}
-}
-
-pub fn print_response<T>(
-    resp: &Result<T, PhylumApiError>,
-    pretty_print: bool,
-    filter: Option<Filter>,
-) where
-    T: std::fmt::Debug + Serialize + Summarize,
-{
-    log::debug!("==> {:?}", resp);
-
-    match resp {
-        Ok(resp) => {
-            if pretty_print {
-                resp.summarize(filter);
-            } else {
-                // Use write! as a workaround to avoid https://github.com/rust-lang/rust/issues/46016
-                //  when piping output to an external program
-                let mut stdout = io::stdout();
-                write!(
-                    &mut stdout,
-                    "{}",
-                    serde_json::to_string_pretty(&resp).unwrap_or_else(|e| {
-                        log::error!("Failed to serialize json response: {}", e);
-                        "".to_string()
-                    })
-                )
-                .unwrap_or_else(|e| log::debug!("Failed writing to stdout: {}", e));
-            }
-        },
-        Err(err) => {
-            print_user_failure!("Response error:\n{}", err);
-        },
-    }
 }
 
 /// Prints a verbose message informing the user that an update is available.
